Centralised location_functions
