+++ Pathogens/Utility/location_utils/__init__.py
@@ -1,2 +1,2 @@
-from .location_util_functions import add_latlng_to_df, get_alternative_names, get_city, get_country_code, \
+from .location_functions import add_latlng_to_df, get_alternative_names, get_city, get_country_code, \
     compute_pin_info
